Fixup Py_ssize_t on Windows VS build
.. Py_ssize_t is not always a long, so lets use the python
   type in our bindings to keep the buffer code simple.

+++ src/Python/SIP/Bindings.h
@@ -1,6 +1,8 @@
 #include <QString>
 #include "RideFile.h"
 
+#include <pyport.h> // for Py_ssize_t
+
 class PythonDataSeries {
 
     public:
@@ -10,7 +12,7 @@ class PythonDataSeries {
         ~PythonDataSeries();
 
         RideFile::SeriesType series;
-        long count;
+        Py_ssize_t count;
         double *data;
 };
